shifts *KeyedVectors to keyedvectors.py
