+++ gensim/models/doc2vec.py
@@ -55,18 +55,17 @@ except ImportError:
 from collections import namedtuple, defaultdict
 from timeit import default_timer
 
-from numpy import zeros, array, float32 as REAL, empty, ones, \
-    memmap as np_memmap, sqrt, newaxis, ndarray, dot, vstack, \
-    divide as np_divide, integer, dtype
+from numpy import zeros, float32 as REAL, empty, ones, \
+    memmap as np_memmap, vstack, integer, dtype
 
 
-from gensim.utils import call_on_class_only, deprecated
+from gensim.utils import call_on_class_only
 from gensim import utils, matutils  # utility fnc for pickling, common scipy operations etc
 from gensim.models.word2vec import Word2VecKeyedVectors, Word2VecVocab, Word2VecTrainables
-from six.moves import xrange, zip
+from six.moves import xrange
 from six import string_types, integer_types
 from gensim.models.base_any2vec import BaseWordEmbedddingsModel
-from gensim.models.keyedvectors import WordEmbeddingsKeyedVectors, BaseKeyedVectors
+from gensim.models.keyedvectors import Doc2VecKeyedVectors
 from types import GeneratorType
 
 logger = logging.getLogger(__name__)
@@ -695,379 +694,6 @@ class Doc2VecTrainables(Word2VecTrainables):
         doctag_vectors[0] = self.seeded_vector(' '.join(doc_words))
         doctag_locks = ones(1, dtype=REAL)
         return doctag_vectors, doctag_locks
-
-
-class Doc2VecKeyedVectors(BaseKeyedVectors):
-    def __init__(self):
-        self.doctags = {}  # string -> Doctag (only filled if necessary)
-        self.max_rawint = -1  # highest rawint-indexed doctag
-        self.offset2doctag = []  # int offset-past-(max_rawint+1) -> String (only filled if necessary)
-        self.count = 0
-        self.vectors_docs = []
-        self.mapfile_path = None
-
-    @property
-    def index2entity(self):
-        return self.offset2doctag
-
-    @index2entity.setter
-    def index2entity(self, value):
-        self.offset2doctag = value
-
-    @property
-    @deprecated("Attribute will be removed in 4.0.0, use docvecs.vectors_docs instead")
-    def doctag_syn0(self):
-        return self.vectors_docs
-
-    @property
-    @deprecated("Attribute will be removed in 4.0.0, use docvecs.vectors_docs_norm instead")
-    def doctag_syn0norm(self):
-        return self.vectors_docs_norm
-
-    def __getitem__(self, index):
-        """
-        Accept a single key (int or string tag) or list of keys as input.
-
-        If a single string or int, return designated tag's vector
-        representation, as a 1D numpy array.
-
-        If a list, return designated tags' vector representations as a
-        2D numpy array: #tags x #vector_size.
-        """
-        if index in self:
-            if isinstance(index, string_types + integer_types + (integer,)):
-                return self.vectors_docs[self._int_index(index, self.doctags, self.max_rawint)]
-            return vstack([self[i] for i in index])
-        raise KeyError("tag '%s' not seen in training corpus/invalid" % index)
-
-    def __contains__(self, index):
-        if isinstance(index, integer_types + (integer,)):
-            return index < self.count
-        else:
-            return index in self.doctags
-
-    def __len__(self):
-        return self.count
-
-    def save(self, *args, **kwargs):
-        """Saves the keyedvectors. This saved model can be loaded again using
-        :func:`~gensim.models.doc2vec.Doc2VecKeyedVectors.load` which supports
-        operations on trained document vectors like `most_similar`.
-
-        Parameters
-        ----------
-        fname : str
-            Path to the file.
-
-        Returns
-        -------
-        None
-
-        """
-        # don't bother storing the cached normalized vectors
-        kwargs['ignore'] = kwargs.get('ignore', ['vectors_docs_norm'])
-        super(Doc2VecKeyedVectors, self).save(*args, **kwargs)
-
-    def init_sims(self, replace=False):
-        """
-        Precompute L2-normalized vectors.
-
-        If `replace` is set, forget the original vectors and only keep the normalized
-        ones = saves lots of memory!
-
-        Note that you **cannot continue training or inference** after doing a replace.
-        The model becomes effectively read-only = you can call `most_similar`, `similarity`
-        etc., but not `train` or `infer_vector`.
-
-        """
-        if getattr(self, 'vectors_docs_norm', None) is None or replace:
-            logger.info("precomputing L2-norms of doc weight vectors")
-            if replace:
-                for i in xrange(self.vectors_docs.shape[0]):
-                    self.vectors_docs[i, :] /= sqrt((self.vectors_docs[i, :] ** 2).sum(-1))
-                self.vectors_docs_norm = self.vectors_docs
-            else:
-                if self.mapfile_path:
-                    self.vectors_docs_norm = np_memmap(
-                        self.mapfile_path + '.vectors_docs_norm', dtype=REAL,
-                        mode='w+', shape=self.vectors_docs.shape)
-                else:
-                    self.vectors_docs_norm = empty(self.vectors_docs.shape, dtype=REAL)
-                np_divide(
-                    self.vectors_docs, sqrt((self.vectors_docs ** 2).sum(-1))[..., newaxis], self.vectors_docs_norm)
-
-    def most_similar(self, positive=None, negative=None, topn=10, clip_start=0, clip_end=None, indexer=None):
-        """
-        Find the top-N most similar docvecs known from training. Positive docs contribute
-        positively towards the similarity, negative docs negatively.
-
-        This method computes cosine similarity between a simple mean of the projection
-        weight vectors of the given docs. Docs may be specified as vectors, integer indexes
-        of trained docvecs, or if the documents were originally presented with string tags,
-        by the corresponding tags.
-
-        The 'clip_start' and 'clip_end' allow limiting results to a particular contiguous
-        range of the underlying `vectors_docs_norm` vectors. (This may be useful if the ordering
-        there was chosen to be significant, such as more popular tag IDs in lower indexes.)
-
-        Parameters
-        ----------
-        positive : :obj: `list`
-            List of Docs specifed as vectors, integer indexes of trained docvecs or string tags
-            that contribute positively.
-        negative : :obj: `list`
-            List of Docs specifed as vectors, integer indexes of trained docvecs or string tags
-            that contribute negatively.
-        topn : int
-            Number of top-N similar docvecs to return.
-        clip_start : int
-            Start clipping index.
-        clip_end : int
-            End clipping index.
-
-        Returns
-        -------
-        :obj: `list` of :obj: `tuple`
-            Returns a list of tuples (doc, similarity)
-
-        """
-        if positive is None:
-            positive = []
-        if negative is None:
-            negative = []
-
-        self.init_sims()
-        clip_end = clip_end or len(self.vectors_docs_norm)
-
-        if isinstance(positive, string_types + integer_types + (integer,)) and not negative:
-            # allow calls like most_similar('dog'), as a shorthand for most_similar(['dog'])
-            positive = [positive]
-
-        # add weights for each doc, if not already present; default to 1.0 for positive and -1.0 for negative docs
-        positive = [
-            (doc, 1.0) if isinstance(doc, string_types + integer_types + (ndarray, integer))
-            else doc for doc in positive
-        ]
-        negative = [
-            (doc, -1.0) if isinstance(doc, string_types + integer_types + (ndarray, integer))
-            else doc for doc in negative
-        ]
-
-        # compute the weighted average of all docs
-        all_docs, mean = set(), []
-        for doc, weight in positive + negative:
-            if isinstance(doc, ndarray):
-                mean.append(weight * doc)
-            elif doc in self.doctags or doc < self.count:
-                mean.append(weight * self.vectors_docs_norm[self._int_index(doc, self.doctags, self.max_rawint)])
-                all_docs.add(self._int_index(doc, self.doctags, self.max_rawint))
-            else:
-                raise KeyError("doc '%s' not in trained set" % doc)
-        if not mean:
-            raise ValueError("cannot compute similarity with no input")
-        mean = matutils.unitvec(array(mean).mean(axis=0)).astype(REAL)
-
-        if indexer is not None:
-            return indexer.most_similar(mean, topn)
-
-        dists = dot(self.vectors_docs_norm[clip_start:clip_end], mean)
-        if not topn:
-            return dists
-        best = matutils.argsort(dists, topn=topn + len(all_docs), reverse=True)
-        # ignore (don't return) docs from the input
-        result = [
-            (self._index_to_doctag(sim + clip_start, self.offset2doctag, self.max_rawint), float(dists[sim]))
-            for sim in best
-            if (sim + clip_start) not in all_docs
-        ]
-        return result[:topn]
-
-    def doesnt_match(self, docs):
-        """
-        Which doc from the given list doesn't go with the others?
-
-        (TODO: Accept vectors of out-of-training-set docs, as if from inference.)
-
-        Parameters
-        ----------
-        docs : :obj: `list` of (str or int)
-            List of seen documents specified by their corresponding string tags or integer indices.
-
-        Returns
-        -------
-        str or int
-            The document further away from the mean of all the documents.
-
-        """
-        self.init_sims()
-
-        docs = [doc for doc in docs if doc in self.doctags or 0 <= doc < self.count]  # filter out unknowns
-        logger.debug("using docs %s", docs)
-        if not docs:
-            raise ValueError("cannot select a doc from an empty list")
-        vectors = vstack(
-            self.vectors_docs_norm[self._int_index(doc, self.doctags, self.max_rawint)] for doc in docs).astype(REAL)
-        mean = matutils.unitvec(vectors.mean(axis=0)).astype(REAL)
-        dists = dot(vectors, mean)
-        return sorted(zip(dists, docs))[0][1]
-
-    def similarity(self, d1, d2):
-        """
-        Compute cosine similarity between two docvecs in the trained set, specified by int index or
-        string tag. (TODO: Accept vectors of out-of-training-set docs, as if from inference.)
-
-        Parameters
-        ----------
-        d1 : int or str
-            Indicate the first document by it's string tag or integer index.
-        d2 : int or str
-            Indicate the second document by it's string tag or integer index.
-
-        Returns
-        -------
-        float
-            The cosine similarity between the vectors of the two documents.
-
-        """
-        return dot(matutils.unitvec(self[d1]), matutils.unitvec(self[d2]))
-
-    def n_similarity(self, ds1, ds2):
-        """
-        Compute cosine similarity between two sets of docvecs from the trained set, specified by int
-        index or string tag. (TODO: Accept vectors of out-of-training-set docs, as if from inference.)
-
-        Parameters
-        ----------
-        ds1 : :obj: `list` of (str or int)
-            Specify the first set of documents as a list of their integer indices or string tags.
-        ds2 : :obj: `list` of (str or int)
-            Specify the second set of documents as a list of their integer indices or string tags.
-
-        Returns
-        -------
-        float
-            The cosine similarity between the means of the documents in each of the two sets.
-
-        """
-        v1 = [self[doc] for doc in ds1]
-        v2 = [self[doc] for doc in ds2]
-        return dot(matutils.unitvec(array(v1).mean(axis=0)), matutils.unitvec(array(v2).mean(axis=0)))
-
-    # required by base keyed vectors class
-    def distances(self, d1, other_docs=()):
-        """Compute distances from given document (string tag or int index) to all documents in `other_docs`.
-        If `other_docs` is empty, return distance between `d1` and all documents seen during training.
-        """
-        input_vector = self[d1]
-        if not other_docs:
-            other_vectors = self.vectors_docs
-        else:
-            other_vectors = self[other_vectors]
-        return 1 - WordEmbeddingsKeyedVectors.cosine_similarities(input_vector, other_vectors)
-
-    def similarity_unseen_docs(self, model, doc_words1, doc_words2, alpha=0.1, min_alpha=0.0001, steps=5):
-        """
-        Compute cosine similarity between two post-bulk out of training documents.
-
-        Parameters
-        ----------
-        model : :obj: `~gensim.models.doc2vec.Doc2Vec`
-            An instance of a trained `Doc2Vec` model.
-        doc_words1 : :obj: `list` of :obj: `str`
-            The first document. Document should be a list of (word) tokens.
-        doc_words2 : :obj: `list` of :obj: `str`
-            The second document. Document should be a list of (word) tokens.
-        alpha : float
-            The initial learning rate.
-        min_alpha : float
-            Learning rate will linearly drop to `min_alpha` as training progresses.
-        steps : int
-            Number of times to train the new document.
-
-        Returns
-        -------
-        float
-            The cosine similarity between the unseen documents.
-
-        """
-        d1 = model.infer_vector(doc_words=doc_words1, alpha=alpha, min_alpha=min_alpha, steps=steps)
-        d2 = model.infer_vector(doc_words=doc_words2, alpha=alpha, min_alpha=min_alpha, steps=steps)
-        return dot(matutils.unitvec(d1), matutils.unitvec(d2))
-
-    def save_word2vec_format(self, fname, prefix='*dt_', fvocab=None,
-                             total_vec=None, binary=False, write_first_line=True):
-        """Store the input-hidden weight matrix in the same format used by the original
-        C word2vec-tool, for compatibility.
-
-        Parameters
-        ----------
-        fname : str
-            The file path used to save the vectors in.
-        prefix : str
-            Uniquely identifies doctags from word vocab, and avoids collision
-            in case of repeated string in doctag and word vocab.
-        fvocab : str
-            Optional file path used to save the vocabulary
-        binary : bool
-            If True, the data wil be saved in binary word2vec format, else it will be saved in plain text.
-        total_vec :  int
-            Optional parameter to explicitly specify total no. of vectors
-            (in case word vectors are appended with document vectors afterwards)
-        write_first_line : bool
-            Whether to print the first line in the file. Useful when saving doc-vectors after word-vectors.
-
-        Returns
-        -------
-        None
-
-        """
-        total_vec = total_vec or len(self)
-        with utils.smart_open(fname, 'ab') as fout:
-            if write_first_line:
-                logger.info("storing %sx%s projection weights into %s", total_vec, self.vectors_docs.shape[1], fname)
-                fout.write(utils.to_utf8("%s %s\n" % (total_vec, self.vectors_docs.shape[1])))
-            # store as in input order
-            for i in range(len(self)):
-                doctag = u"%s%s" % (prefix, self._index_to_doctag(i, self.offset2doctag, self.max_rawint))
-                row = self.vectors_docs[i]
-                if binary:
-                    fout.write(utils.to_utf8(doctag) + b" " + row.tostring())
-                else:
-                    fout.write(utils.to_utf8("%s %s\n" % (doctag, ' '.join("%f" % val for val in row))))
-
-    @staticmethod
-    def _int_index(index, doctags, max_rawint):
-        """Return int index for either string or int index"""
-        if isinstance(index, integer_types + (integer,)):
-            return index
-        else:
-            return max_rawint + 1 + doctags[index].offset
-
-    @staticmethod
-    def _index_to_doctag(i_index, offset2doctag, max_rawint):
-        """Return string key for given i_index, if available. Otherwise return raw int doctag (same int)."""
-        candidate_offset = i_index - max_rawint - 1
-        if 0 <= candidate_offset < len(offset2doctag):
-            return offset2doctag[candidate_offset]
-        else:
-            return i_index
-
-    # for backward compatibility
-    def index_to_doctag(self, i_index):
-        """Return string key for given i_index, if available. Otherwise return raw int doctag (same int)."""
-        candidate_offset = i_index - self.max_rawint - 1
-        if 0 <= candidate_offset < len(self.offset2doctag):
-            return self.ffset2doctag[candidate_offset]
-        else:
-            return i_index
-
-    # for backward compatibility
-    def int_index(index, doctags, max_rawint):
-        """Return int index for either string or int index"""
-        if isinstance(index, integer_types + (integer,)):
-            return index
-        else:
-            return max_rawint + 1 + doctags[index].offset
 
 
 class TaggedBrownCorpus(object):
