+++ gensim/models/keyedvectors.py
@@ -71,15 +71,17 @@ try:
 except ImportError:
     PYEMD_EXT = False
 
-from numpy import dot, zeros, float32 as REAL,\
-    double, array, vstack, sqrt, newaxis,\
-    ndarray, sum as np_sum, prod, argmax
+from numpy import dot, zeros, float32 as REAL, empty, memmap as np_memmap, \
+    double, array, vstack, sqrt, newaxis, integer, \
+    ndarray, sum as np_sum, prod, argmax, divide as np_divide
 import numpy as np
 from gensim import utils, matutils  # utility fnc for pickling, common scipy operations etc
 from gensim.corpora.dictionary import Dictionary
-from six import string_types
-from six.moves import xrange
+from six import string_types, integer_types
+from six.moves import xrange, zip
 from scipy import stats
+from gensim.utils import deprecated
+from gensim.models.utils_any2vec import save_word2vec_format, load_word2vec_format, compute_ngrams
 
 logger = logging.getLogger(__name__)
 
@@ -926,3 +928,628 @@ class WordEmbeddingsKeyedVectors(BaseKeyedVectors):
                 self.vectors_norm = self.vectors
             else:
                 self.vectors_norm = (self.vectors / sqrt((self.vectors ** 2).sum(-1))[..., newaxis]).astype(REAL)
+
+
+class Word2VecKeyedVectors(WordEmbeddingsKeyedVectors):
+    """Class to contain vectors and vocab for the Word2Vec training class and other w2v methods not directly
+    involved in training such as most_similar()
+    """
+    def save_word2vec_format(self, fname, fvocab=None, binary=False, total_vec=None):
+        """Store the input-hidden weight matrix in the same format used by the original
+        C word2vec-tool, for compatibility.
+
+        Parameters
+        ----------
+        fname : str
+            The file path used to save the vectors in
+        fvocab : str
+            Optional file path used to save the vocabulary
+        binary : bool
+            If True, the data wil be saved in binary word2vec format, else it will be saved in plain text.
+        total_vec :  int
+            Optional parameter to explicitly specify total no. of vectors
+            (in case word vectors are appended with document vectors afterwards)
+
+        Returns
+        -------
+        None
+
+        """
+        # from gensim.models.word2vec import save_word2vec_format
+        save_word2vec_format(
+            fname, self.vocab, self.vectors, fvocab=fvocab, binary=binary, total_vec=total_vec)
+
+    @classmethod
+    def load_word2vec_format(cls, fname, fvocab=None, binary=False, encoding='utf8', unicode_errors='strict',
+                             limit=None, datatype=REAL):
+        """Load the input-hidden weight matrix from the original C word2vec-tool format.
+
+        Note that the information stored in the file is incomplete (the binary tree is missing),
+        so while you can query for word similarity etc., you cannot continue training
+        with a model loaded this way.
+
+        Parameters
+        ----------
+        fname : str
+            The file path to the saved word2vec-format file.
+        fvocab : str
+                Optional file path to the vocabulary.Word counts are read from `fvocab` filename,
+                if set (this is the file generated by `-save-vocab` flag of the original C tool).
+        binary : bool
+            If True, indicates whether the data is in binary word2vec format.
+        encoding : str
+            If you trained the C model using non-utf8 encoding for words, specify that
+            encoding in `encoding`.
+        unicode_errors : str
+            default 'strict', is a string suitable to be passed as the `errors`
+            argument to the unicode() (Python 2.x) or str() (Python 3.x) function. If your source
+            file may include word tokens truncated in the middle of a multibyte unicode character
+            (as is common from the original word2vec.c tool), 'ignore' or 'replace' may help.
+        limit : int
+            Sets a maximum number of word-vectors to read from the file. The default,
+            None, means read all.
+        datatype : :class: `numpy.float*`
+            (Experimental) Can coerce dimensions to a non-default float type (such
+            as np.float16) to save memory. (Such types may result in much slower bulk operations
+            or incompatibility with optimized routines.)
+
+        Returns
+        -------
+        :obj: `~gensim.models.word2vec.Wod2Vec`
+            Returns the loaded model as an instance of :class: `~gensim.models.word2vec.Wod2Vec`.
+
+        """
+        # from gensim.models.word2vec import load_word2vec_format
+        return load_word2vec_format(
+            Word2VecKeyedVectors, fname, fvocab=fvocab, binary=binary, encoding=encoding, unicode_errors=unicode_errors,
+            limit=limit, datatype=datatype)
+
+    def get_keras_embedding(self, train_embeddings=False):
+        """Return a Keras 'Embedding' layer with weights set as the Word2Vec model's learned word embeddings
+
+        Parameter
+        ---------
+        train_embeddings : bool
+            If False, the weights are frozen and stopped from being updated.
+            If True, the weights can/will be further trained/updated.
+
+        Return
+        ------
+        :obj: `keras.layers.Embedding`
+            Embedding layer
+
+        """
+        try:
+            from keras.layers import Embedding
+        except ImportError:
+            raise ImportError("Please install Keras to use this function")
+        weights = self.vectors
+
+        # set `trainable` as `False` to use the pretrained word embedding
+        # No extra mem usage here as `Embedding` layer doesn't create any new matrix for weights
+        layer = Embedding(
+            input_dim=weights.shape[0], output_dim=weights.shape[1],
+            weights=[weights], trainable=train_embeddings
+        )
+        return layer
+KeyedVectors = Word2VecKeyedVectors  # alias for backward compatibility
+
+
+class Doc2VecKeyedVectors(BaseKeyedVectors):
+    def __init__(self):
+        self.doctags = {}  # string -> Doctag (only filled if necessary)
+        self.max_rawint = -1  # highest rawint-indexed doctag
+        self.offset2doctag = []  # int offset-past-(max_rawint+1) -> String (only filled if necessary)
+        self.count = 0
+        self.vectors_docs = []
+        self.mapfile_path = None
+
+    @property
+    def index2entity(self):
+        return self.offset2doctag
+
+    @index2entity.setter
+    def index2entity(self, value):
+        self.offset2doctag = value
+
+    @property
+    @deprecated("Attribute will be removed in 4.0.0, use docvecs.vectors_docs instead")
+    def doctag_syn0(self):
+        return self.vectors_docs
+
+    @property
+    @deprecated("Attribute will be removed in 4.0.0, use docvecs.vectors_docs_norm instead")
+    def doctag_syn0norm(self):
+        return self.vectors_docs_norm
+
+    def __getitem__(self, index):
+        """
+        Accept a single key (int or string tag) or list of keys as input.
+
+        If a single string or int, return designated tag's vector
+        representation, as a 1D numpy array.
+
+        If a list, return designated tags' vector representations as a
+        2D numpy array: #tags x #vector_size.
+        """
+        if index in self:
+            if isinstance(index, string_types + integer_types + (integer,)):
+                return self.vectors_docs[self._int_index(index, self.doctags, self.max_rawint)]
+            return vstack([self[i] for i in index])
+        raise KeyError("tag '%s' not seen in training corpus/invalid" % index)
+
+    def __contains__(self, index):
+        if isinstance(index, integer_types + (integer,)):
+            return index < self.count
+        else:
+            return index in self.doctags
+
+    def __len__(self):
+        return self.count
+
+    def save(self, *args, **kwargs):
+        """Saves the keyedvectors. This saved model can be loaded again using
+        :func:`~gensim.models.doc2vec.Doc2VecKeyedVectors.load` which supports
+        operations on trained document vectors like `most_similar`.
+
+        Parameters
+        ----------
+        fname : str
+            Path to the file.
+
+        Returns
+        -------
+        None
+
+        """
+        # don't bother storing the cached normalized vectors
+        kwargs['ignore'] = kwargs.get('ignore', ['vectors_docs_norm'])
+        super(Doc2VecKeyedVectors, self).save(*args, **kwargs)
+
+    def init_sims(self, replace=False):
+        """
+        Precompute L2-normalized vectors.
+
+        If `replace` is set, forget the original vectors and only keep the normalized
+        ones = saves lots of memory!
+
+        Note that you **cannot continue training or inference** after doing a replace.
+        The model becomes effectively read-only = you can call `most_similar`, `similarity`
+        etc., but not `train` or `infer_vector`.
+
+        """
+        if getattr(self, 'vectors_docs_norm', None) is None or replace:
+            logger.info("precomputing L2-norms of doc weight vectors")
+            if replace:
+                for i in xrange(self.vectors_docs.shape[0]):
+                    self.vectors_docs[i, :] /= sqrt((self.vectors_docs[i, :] ** 2).sum(-1))
+                self.vectors_docs_norm = self.vectors_docs
+            else:
+                if self.mapfile_path:
+                    self.vectors_docs_norm = np_memmap(
+                        self.mapfile_path + '.vectors_docs_norm', dtype=REAL,
+                        mode='w+', shape=self.vectors_docs.shape)
+                else:
+                    self.vectors_docs_norm = empty(self.vectors_docs.shape, dtype=REAL)
+                np_divide(
+                    self.vectors_docs, sqrt((self.vectors_docs ** 2).sum(-1))[..., newaxis], self.vectors_docs_norm)
+
+    def most_similar(self, positive=None, negative=None, topn=10, clip_start=0, clip_end=None, indexer=None):
+        """
+        Find the top-N most similar docvecs known from training. Positive docs contribute
+        positively towards the similarity, negative docs negatively.
+
+        This method computes cosine similarity between a simple mean of the projection
+        weight vectors of the given docs. Docs may be specified as vectors, integer indexes
+        of trained docvecs, or if the documents were originally presented with string tags,
+        by the corresponding tags.
+
+        The 'clip_start' and 'clip_end' allow limiting results to a particular contiguous
+        range of the underlying `vectors_docs_norm` vectors. (This may be useful if the ordering
+        there was chosen to be significant, such as more popular tag IDs in lower indexes.)
+
+        Parameters
+        ----------
+        positive : :obj: `list`
+            List of Docs specifed as vectors, integer indexes of trained docvecs or string tags
+            that contribute positively.
+        negative : :obj: `list`
+            List of Docs specifed as vectors, integer indexes of trained docvecs or string tags
+            that contribute negatively.
+        topn : int
+            Number of top-N similar docvecs to return.
+        clip_start : int
+            Start clipping index.
+        clip_end : int
+            End clipping index.
+
+        Returns
+        -------
+        :obj: `list` of :obj: `tuple`
+            Returns a list of tuples (doc, similarity)
+
+        """
+        if positive is None:
+            positive = []
+        if negative is None:
+            negative = []
+
+        self.init_sims()
+        clip_end = clip_end or len(self.vectors_docs_norm)
+
+        if isinstance(positive, string_types + integer_types + (integer,)) and not negative:
+            # allow calls like most_similar('dog'), as a shorthand for most_similar(['dog'])
+            positive = [positive]
+
+        # add weights for each doc, if not already present; default to 1.0 for positive and -1.0 for negative docs
+        positive = [
+            (doc, 1.0) if isinstance(doc, string_types + integer_types + (ndarray, integer))
+            else doc for doc in positive
+        ]
+        negative = [
+            (doc, -1.0) if isinstance(doc, string_types + integer_types + (ndarray, integer))
+            else doc for doc in negative
+        ]
+
+        # compute the weighted average of all docs
+        all_docs, mean = set(), []
+        for doc, weight in positive + negative:
+            if isinstance(doc, ndarray):
+                mean.append(weight * doc)
+            elif doc in self.doctags or doc < self.count:
+                mean.append(weight * self.vectors_docs_norm[self._int_index(doc, self.doctags, self.max_rawint)])
+                all_docs.add(self._int_index(doc, self.doctags, self.max_rawint))
+            else:
+                raise KeyError("doc '%s' not in trained set" % doc)
+        if not mean:
+            raise ValueError("cannot compute similarity with no input")
+        mean = matutils.unitvec(array(mean).mean(axis=0)).astype(REAL)
+
+        if indexer is not None:
+            return indexer.most_similar(mean, topn)
+
+        dists = dot(self.vectors_docs_norm[clip_start:clip_end], mean)
+        if not topn:
+            return dists
+        best = matutils.argsort(dists, topn=topn + len(all_docs), reverse=True)
+        # ignore (don't return) docs from the input
+        result = [
+            (self._index_to_doctag(sim + clip_start, self.offset2doctag, self.max_rawint), float(dists[sim]))
+            for sim in best
+            if (sim + clip_start) not in all_docs
+        ]
+        return result[:topn]
+
+    def doesnt_match(self, docs):
+        """
+        Which doc from the given list doesn't go with the others?
+
+        (TODO: Accept vectors of out-of-training-set docs, as if from inference.)
+
+        Parameters
+        ----------
+        docs : :obj: `list` of (str or int)
+            List of seen documents specified by their corresponding string tags or integer indices.
+
+        Returns
+        -------
+        str or int
+            The document further away from the mean of all the documents.
+
+        """
+        self.init_sims()
+
+        docs = [doc for doc in docs if doc in self.doctags or 0 <= doc < self.count]  # filter out unknowns
+        logger.debug("using docs %s", docs)
+        if not docs:
+            raise ValueError("cannot select a doc from an empty list")
+        vectors = vstack(
+            self.vectors_docs_norm[self._int_index(doc, self.doctags, self.max_rawint)] for doc in docs).astype(REAL)
+        mean = matutils.unitvec(vectors.mean(axis=0)).astype(REAL)
+        dists = dot(vectors, mean)
+        return sorted(zip(dists, docs))[0][1]
+
+    def similarity(self, d1, d2):
+        """
+        Compute cosine similarity between two docvecs in the trained set, specified by int index or
+        string tag. (TODO: Accept vectors of out-of-training-set docs, as if from inference.)
+
+        Parameters
+        ----------
+        d1 : int or str
+            Indicate the first document by it's string tag or integer index.
+        d2 : int or str
+            Indicate the second document by it's string tag or integer index.
+
+        Returns
+        -------
+        float
+            The cosine similarity between the vectors of the two documents.
+
+        """
+        return dot(matutils.unitvec(self[d1]), matutils.unitvec(self[d2]))
+
+    def n_similarity(self, ds1, ds2):
+        """
+        Compute cosine similarity between two sets of docvecs from the trained set, specified by int
+        index or string tag. (TODO: Accept vectors of out-of-training-set docs, as if from inference.)
+
+        Parameters
+        ----------
+        ds1 : :obj: `list` of (str or int)
+            Specify the first set of documents as a list of their integer indices or string tags.
+        ds2 : :obj: `list` of (str or int)
+            Specify the second set of documents as a list of their integer indices or string tags.
+
+        Returns
+        -------
+        float
+            The cosine similarity between the means of the documents in each of the two sets.
+
+        """
+        v1 = [self[doc] for doc in ds1]
+        v2 = [self[doc] for doc in ds2]
+        return dot(matutils.unitvec(array(v1).mean(axis=0)), matutils.unitvec(array(v2).mean(axis=0)))
+
+    # required by base keyed vectors class
+    def distances(self, d1, other_docs=()):
+        """Compute distances from given document (string tag or int index) to all documents in `other_docs`.
+        If `other_docs` is empty, return distance between `d1` and all documents seen during training.
+        """
+        input_vector = self[d1]
+        if not other_docs:
+            other_vectors = self.vectors_docs
+        else:
+            other_vectors = self[other_vectors]
+        return 1 - WordEmbeddingsKeyedVectors.cosine_similarities(input_vector, other_vectors)
+
+    def similarity_unseen_docs(self, model, doc_words1, doc_words2, alpha=0.1, min_alpha=0.0001, steps=5):
+        """
+        Compute cosine similarity between two post-bulk out of training documents.
+
+        Parameters
+        ----------
+        model : :obj: `~gensim.models.doc2vec.Doc2Vec`
+            An instance of a trained `Doc2Vec` model.
+        doc_words1 : :obj: `list` of :obj: `str`
+            The first document. Document should be a list of (word) tokens.
+        doc_words2 : :obj: `list` of :obj: `str`
+            The second document. Document should be a list of (word) tokens.
+        alpha : float
+            The initial learning rate.
+        min_alpha : float
+            Learning rate will linearly drop to `min_alpha` as training progresses.
+        steps : int
+            Number of times to train the new document.
+
+        Returns
+        -------
+        float
+            The cosine similarity between the unseen documents.
+
+        """
+        d1 = model.infer_vector(doc_words=doc_words1, alpha=alpha, min_alpha=min_alpha, steps=steps)
+        d2 = model.infer_vector(doc_words=doc_words2, alpha=alpha, min_alpha=min_alpha, steps=steps)
+        return dot(matutils.unitvec(d1), matutils.unitvec(d2))
+
+    def save_word2vec_format(self, fname, prefix='*dt_', fvocab=None,
+                             total_vec=None, binary=False, write_first_line=True):
+        """Store the input-hidden weight matrix in the same format used by the original
+        C word2vec-tool, for compatibility.
+
+        Parameters
+        ----------
+        fname : str
+            The file path used to save the vectors in.
+        prefix : str
+            Uniquely identifies doctags from word vocab, and avoids collision
+            in case of repeated string in doctag and word vocab.
+        fvocab : str
+            Optional file path used to save the vocabulary
+        binary : bool
+            If True, the data wil be saved in binary word2vec format, else it will be saved in plain text.
+        total_vec :  int
+            Optional parameter to explicitly specify total no. of vectors
+            (in case word vectors are appended with document vectors afterwards)
+        write_first_line : bool
+            Whether to print the first line in the file. Useful when saving doc-vectors after word-vectors.
+
+        Returns
+        -------
+        None
+
+        """
+        total_vec = total_vec or len(self)
+        with utils.smart_open(fname, 'ab') as fout:
+            if write_first_line:
+                logger.info("storing %sx%s projection weights into %s", total_vec, self.vectors_docs.shape[1], fname)
+                fout.write(utils.to_utf8("%s %s\n" % (total_vec, self.vectors_docs.shape[1])))
+            # store as in input order
+            for i in range(len(self)):
+                doctag = u"%s%s" % (prefix, self._index_to_doctag(i, self.offset2doctag, self.max_rawint))
+                row = self.vectors_docs[i]
+                if binary:
+                    fout.write(utils.to_utf8(doctag) + b" " + row.tostring())
+                else:
+                    fout.write(utils.to_utf8("%s %s\n" % (doctag, ' '.join("%f" % val for val in row))))
+
+    @staticmethod
+    def _int_index(index, doctags, max_rawint):
+        """Return int index for either string or int index"""
+        if isinstance(index, integer_types + (integer,)):
+            return index
+        else:
+            return max_rawint + 1 + doctags[index].offset
+
+    @staticmethod
+    def _index_to_doctag(i_index, offset2doctag, max_rawint):
+        """Return string key for given i_index, if available. Otherwise return raw int doctag (same int)."""
+        candidate_offset = i_index - max_rawint - 1
+        if 0 <= candidate_offset < len(offset2doctag):
+            return offset2doctag[candidate_offset]
+        else:
+            return i_index
+
+    # for backward compatibility
+    def index_to_doctag(self, i_index):
+        """Return string key for given i_index, if available. Otherwise return raw int doctag (same int)."""
+        candidate_offset = i_index - self.max_rawint - 1
+        if 0 <= candidate_offset < len(self.offset2doctag):
+            return self.ffset2doctag[candidate_offset]
+        else:
+            return i_index
+
+    # for backward compatibility
+    def int_index(index, doctags, max_rawint):
+        """Return int index for either string or int index"""
+        if isinstance(index, integer_types + (integer,)):
+            return index
+        else:
+            return max_rawint + 1 + doctags[index].offset
+
+
+class FastTextKeyedVectors(WordEmbeddingsKeyedVectors):
+    """
+    Class to contain vectors and vocab for the FastText training class and other methods not directly
+    involved in training such as most_similar()
+    """
+
+    def __init__(self):
+        super(FastTextKeyedVectors, self).__init__()
+        self.vectors_vocab = None
+        self.vectors_vocab_norm = None
+        self.vectors_ngrams = None
+        self.vectors_ngrams_norm = None
+        self.ngrams = {}
+        self.hash2index = {}
+        self.ngrams_word = {}
+        self.min_n = 0
+        self.max_n = 0
+
+    @property
+    @deprecated("Attribute will be removed in 4.0.0, use self.wv.vectors_vocab instead")
+    def syn0_vocab(self):
+        return self.vectors_vocab
+
+    @property
+    @deprecated("Attribute will be removed in 4.0.0, use self.wv.vectors_vocab_norm instead")
+    def syn0_vocab_norm(self):
+        return self.vectors_vocab_norm
+
+    @property
+    @deprecated("Attribute will be removed in 4.0.0, use self.wv.vectors_ngrams instead")
+    def syn0_ngrams(self):
+        return self.vectors_ngrams
+
+    @property
+    @deprecated("Attribute will be removed in 4.0.0, use self.wv.vectors_ngrams_norm instead")
+    def syn0_ngrams_norm(self):
+        return self.vectors_ngrams_norm
+
+    @property
+    def num_ngram_vectors(self):
+        return self.trainables.num_ngram_vectors
+
+    def __contains__(self, word):
+        """
+        Check if `word` or any character ngrams in `word` are present in the vocabulary.
+        A vector for the word is guaranteed to exist if `__contains__` returns True.
+        """
+        if word in self.vocab:
+            return True
+        else:
+            # from gensim.models.fasttext import compute_ngrams
+            char_ngrams = compute_ngrams(word, self.min_n, self.max_n)
+            return any(ng in self.ngrams for ng in char_ngrams)
+
+    def save(self, *args, **kwargs):
+        """Saves the keyedvectors. This saved model can be loaded again using
+        :func:`~gensim.models.fasttext.FastTextKeyedVectors.load` which supports
+        getting vectors for out-of-vocabulary words.
+
+        Parameters
+        ----------
+        fname : str
+            Path to the file.
+
+        Returns
+        -------
+        None
+
+        """
+        # don't bother storing the cached normalized vectors
+        kwargs['ignore'] = kwargs.get('ignore', ['vectors_norm', 'vectors_vocab_norm', 'vectors_ngrams_norm'])
+        super(FastTextKeyedVectors, self).save(*args, **kwargs)
+
+    def word_vec(self, word, use_norm=False):
+        """
+        Accept a single word as input.
+        Returns the word's representations in vector space, as a 1D numpy array.
+
+        If `use_norm` is True, returns the normalized word vector.
+
+        """
+        if word in self.vocab:
+            return super(FastTextKeyedVectors, self).word_vec(word, use_norm)
+        else:
+            # from gensim.models.fasttext import compute_ngrams
+            word_vec = np.zeros(self.vectors_ngrams.shape[1], dtype=np.float32)
+            ngrams = compute_ngrams(word, self.min_n, self.max_n)
+            ngrams = [ng for ng in ngrams if ng in self.ngrams]
+            if use_norm:
+                ngram_weights = self.vectors_ngrams_norm
+            else:
+                ngram_weights = self.vectors_ngrams
+            for ngram in ngrams:
+                word_vec += ngram_weights[self.ngrams[ngram]]
+            if word_vec.any():
+                return word_vec / len(ngrams)
+            else:  # No ngrams of the word are present in self.ngrams
+                raise KeyError('all ngrams for word %s absent from model' % word)
+
+    def init_sims(self, replace=False):
+        """
+        Precompute L2-normalized vectors.
+
+        If `replace` is set, forget the original vectors and only keep the normalized
+        ones = saves lots of memory!
+
+        Note that you **cannot continue training** after doing a replace. The model becomes
+        effectively read-only = you can only call `most_similar`, `similarity` etc.
+
+        """
+        super(FastTextKeyedVectors, self).init_sims(replace)
+        if getattr(self, 'vectors_ngrams_norm', None) is None or replace:
+            logger.info("precomputing L2-norms of ngram weight vectors")
+            if replace:
+                for i in range(self.vectors_ngrams.shape[0]):
+                    self.vectors_ngrams[i, :] /= sqrt((self.vectors_ngrams[i, :] ** 2).sum(-1))
+                self.vectors_ngrams_norm = self.vectors_ngrams
+            else:
+                self.vectors_ngrams_norm = \
+                    (self.vectors_ngrams / sqrt((self.vectors_ngrams ** 2).sum(-1))[..., newaxis]).astype(REAL)
+
+    def save_word2vec_format(self, fname, fvocab=None, binary=False, total_vec=None):
+        """Store the input-hidden weight matrix in the same format used by the original
+        C word2vec-tool, for compatibility.
+
+        Parameters
+        ----------
+        fname : str
+            The file path used to save the vectors in.
+        fvocab : str
+            Optional file path used to save the vocabulary.
+        binary : bool
+            If True, the data wil be saved in binary word2vec format, else it will be saved in plain text.
+        total_vec :  int
+            Optional parameter to explicitly specify total no. of vectors
+            (in case word vectors are appended with document vectors afterwards).
+
+        Returns
+        -------
+        None
+
+        """
+        # from gensim.models.word2vec import save_word2vec_format
+        save_word2vec_format(
+            fname, self.vocab, self.vectors, fvocab=fvocab, binary=binary, total_vec=total_vec)
