+++ gensim/models/word2vec.py
@@ -110,7 +110,7 @@ import itertools
 import warnings
 
 from gensim.utils import keep_vocab_item, call_on_class_only
-from gensim.models.keyedvectors import WordEmbeddingsKeyedVectors, Vocab
+from gensim.models.keyedvectors import Vocab, Word2VecKeyedVectors
 from gensim.models.base_any2vec import BaseWordEmbedddingsModel,\
     BaseVocabBuilder, BaseModelTrainables
 
@@ -120,7 +120,7 @@ except ImportError:
     from Queue import Queue, Empty
 
 from numpy import exp, dot, zeros, random, dtype, float32 as REAL,\
-    uint32, seterr, array, ascontiguousarray, uint8, vstack, fromstring, sqrt,\
+    uint32, seterr, array, uint8, vstack, fromstring, sqrt,\
     empty, sum as np_sum, ones
 
 from gensim import utils, matutils  # utility fnc for pickling, common scipy operations etc
@@ -1178,265 +1178,6 @@ class Word2VecTrainables(BaseModelTrainables):
 
         # do not suppress learning for already learned words
         self.vectors_lockf = ones(len(vocabulary.vocab), dtype=REAL)  # zeros suppress learning
-
-
-class Word2VecKeyedVectors(WordEmbeddingsKeyedVectors):
-    """Class to contain vectors and vocab for the Word2Vec training class and other w2v methods not directly
-    involved in training such as most_similar()
-    """
-    def save_word2vec_format(self, fname, fvocab=None, binary=False, total_vec=None):
-        """Store the input-hidden weight matrix in the same format used by the original
-        C word2vec-tool, for compatibility.
-
-        Parameters
-        ----------
-        fname : str
-            The file path used to save the vectors in
-        fvocab : str
-            Optional file path used to save the vocabulary
-        binary : bool
-            If True, the data wil be saved in binary word2vec format, else it will be saved in plain text.
-        total_vec :  int
-            Optional parameter to explicitly specify total no. of vectors
-            (in case word vectors are appended with document vectors afterwards)
-
-        Returns
-        -------
-        None
-
-        """
-        save_word2vec_format(fname, self.vocab, self.vectors, fvocab=fvocab, binary=binary, total_vec=total_vec)
-
-    @classmethod
-    def load_word2vec_format(cls, fname, fvocab=None, binary=False, encoding='utf8', unicode_errors='strict',
-                             limit=None, datatype=REAL):
-        """Load the input-hidden weight matrix from the original C word2vec-tool format.
-
-        Note that the information stored in the file is incomplete (the binary tree is missing),
-        so while you can query for word similarity etc., you cannot continue training
-        with a model loaded this way.
-
-        Parameters
-        ----------
-        fname : str
-            The file path to the saved word2vec-format file.
-        fvocab : str
-                Optional file path to the vocabulary.Word counts are read from `fvocab` filename,
-                if set (this is the file generated by `-save-vocab` flag of the original C tool).
-        binary : bool
-            If True, indicates whether the data is in binary word2vec format.
-        encoding : str
-            If you trained the C model using non-utf8 encoding for words, specify that
-            encoding in `encoding`.
-        unicode_errors : str
-            default 'strict', is a string suitable to be passed as the `errors`
-            argument to the unicode() (Python 2.x) or str() (Python 3.x) function. If your source
-            file may include word tokens truncated in the middle of a multibyte unicode character
-            (as is common from the original word2vec.c tool), 'ignore' or 'replace' may help.
-        limit : int
-            Sets a maximum number of word-vectors to read from the file. The default,
-            None, means read all.
-        datatype : :class: `numpy.float*`
-            (Experimental) Can coerce dimensions to a non-default float type (such
-            as np.float16) to save memory. (Such types may result in much slower bulk operations
-            or incompatibility with optimized routines.)
-
-        Returns
-        -------
-        :obj: `~gensim.models.word2vec.Wod2Vec`
-            Returns the loaded model as an instance of :class: `~gensim.models.word2vec.Wod2Vec`.
-
-        """
-        return load_word2vec_format(
-            Word2VecKeyedVectors, fname, fvocab=fvocab, binary=binary, encoding=encoding, unicode_errors=unicode_errors,
-            limit=limit, datatype=datatype)
-
-    def get_keras_embedding(self, train_embeddings=False):
-        """Return a Keras 'Embedding' layer with weights set as the Word2Vec model's learned word embeddings
-
-        Parameter
-        ---------
-        train_embeddings : bool
-            If False, the weights are frozen and stopped from being updated.
-            If True, the weights can/will be further trained/updated.
-
-        Return
-        ------
-        :obj: `keras.layers.Embedding`
-            Embedding layer
-
-        """
-        try:
-            from keras.layers import Embedding
-        except ImportError:
-            raise ImportError("Please install Keras to use this function")
-        weights = self.vectors
-
-        # set `trainable` as `False` to use the pretrained word embedding
-        # No extra mem usage here as `Embedding` layer doesn't create any new matrix for weights
-        layer = Embedding(
-            input_dim=weights.shape[0], output_dim=weights.shape[1],
-            weights=[weights], trainable=train_embeddings
-        )
-        return layer
-
-
-def save_word2vec_format(fname, vocab, vectors, fvocab=None, binary=False, total_vec=None):
-        """Store the input-hidden weight matrix in the same format used by the original
-        C word2vec-tool, for compatibility.
-
-        Parameters
-        ----------
-        fname : str
-            The file path used to save the vectors in
-        fvocab : str
-            Optional file path used to save the vocabulary
-        binary : bool
-            If True, the data wil be saved in binary word2vec format, else it will be saved in plain text.
-        total_vec :  int
-            Optional parameter to explicitly specify total no. of vectors
-            (in case word vectors are appended with document vectors afterwards)
-
-        Returns
-        -------
-        None
-
-        """
-        if not (vocab or vectors):
-            raise RuntimeError("no input")
-        if total_vec is None:
-            total_vec = len(vocab)
-        vector_size = vectors.shape[1]
-        if fvocab is not None:
-            logger.info("storing vocabulary in %s", fvocab)
-            with utils.smart_open(fvocab, 'wb') as vout:
-                for word, vocab_ in sorted(iteritems(vocab), key=lambda item: -item[1].count):
-                    vout.write(utils.to_utf8("%s %s\n" % (word, vocab_.count)))
-        logger.info("storing %sx%s projection weights into %s", total_vec, vector_size, fname)
-        assert (len(vocab), vector_size) == vectors.shape
-        with utils.smart_open(fname, 'wb') as fout:
-            fout.write(utils.to_utf8("%s %s\n" % (total_vec, vector_size)))
-            # store in sorted order: most frequent words at the top
-            for word, vocab_ in sorted(iteritems(vocab), key=lambda item: -item[1].count):
-                row = vectors[vocab_.index]
-                if binary:
-                    fout.write(utils.to_utf8(word) + b" " + row.tostring())
-                else:
-                    fout.write(utils.to_utf8("%s %s\n" % (word, ' '.join("%f" % val for val in row))))
-
-
-def load_word2vec_format(cls, fname, fvocab=None, binary=False, encoding='utf8', unicode_errors='strict',
-                             limit=None, datatype=REAL):
-    """Load the input-hidden weight matrix from the original C word2vec-tool format.
-
-    Note that the information stored in the file is incomplete (the binary tree is missing),
-    so while you can query for word similarity etc., you cannot continue training
-    with a model loaded this way.
-
-    Parameters
-    ----------
-    fname : str
-        The file path to the saved word2vec-format file.
-    fvocab : str
-            Optional file path to the vocabulary.Word counts are read from `fvocab` filename,
-            if set (this is the file generated by `-save-vocab` flag of the original C tool).
-    binary : bool
-        If True, indicates whether the data is in binary word2vec format.
-    encoding : str
-        If you trained the C model using non-utf8 encoding for words, specify that
-        encoding in `encoding`.
-    unicode_errors : str
-        default 'strict', is a string suitable to be passed as the `errors`
-        argument to the unicode() (Python 2.x) or str() (Python 3.x) function. If your source
-        file may include word tokens truncated in the middle of a multibyte unicode character
-        (as is common from the original word2vec.c tool), 'ignore' or 'replace' may help.
-    limit : int
-        Sets a maximum number of word-vectors to read from the file. The default,
-        None, means read all.
-    datatype : :class: `numpy.float*`
-        (Experimental) Can coerce dimensions to a non-default float type (such
-        as np.float16) to save memory. (Such types may result in much slower bulk operations
-        or incompatibility with optimized routines.)
-
-    Returns
-    -------
-    :obj: `cls`
-        Returns the loaded model as an instance of :class: `cls`.
-
-    """
-    counts = None
-    if fvocab is not None:
-        logger.info("loading word counts from %s", fvocab)
-        counts = {}
-        with utils.smart_open(fvocab) as fin:
-            for line in fin:
-                word, count = utils.to_unicode(line).strip().split()
-                counts[word] = int(count)
-
-    logger.info("loading projection weights from %s", fname)
-    with utils.smart_open(fname) as fin:
-        header = utils.to_unicode(fin.readline(), encoding=encoding)
-        vocab_size, vector_size = (int(x) for x in header.split())  # throws for invalid file format
-        if limit:
-            vocab_size = min(vocab_size, limit)
-        result = cls()
-        result.vector_size = vector_size
-        result.vectors = zeros((vocab_size, vector_size), dtype=datatype)
-
-        def add_word(word, weights):
-            word_id = len(result.vocab)
-            if word in result.vocab:
-                logger.warning("duplicate word '%s' in %s, ignoring all but first", word, fname)
-                return
-            if counts is None:
-                # most common scenario: no vocab file given. just make up some bogus counts, in descending order
-                result.vocab[word] = Vocab(index=word_id, count=vocab_size - word_id)
-            elif word in counts:
-                # use count from the vocab file
-                result.vocab[word] = Vocab(index=word_id, count=counts[word])
-            else:
-                # vocab file given, but word is missing -- set count to None (TODO: or raise?)
-                logger.warning("vocabulary file is incomplete: '%s' is missing", word)
-                result.vocab[word] = Vocab(index=word_id, count=None)
-            result.vectors[word_id] = weights
-            result.index2word.append(word)
-
-        if binary:
-            binary_len = dtype(REAL).itemsize * vector_size
-            for _ in xrange(vocab_size):
-                # mixed text and binary: read text first, then binary
-                word = []
-                while True:
-                    ch = fin.read(1)
-                    if ch == b' ':
-                        break
-                    if ch == b'':
-                        raise EOFError("unexpected end of input; is count incorrect or file otherwise damaged?")
-                    if ch != b'\n':  # ignore newlines in front of words (some binary files have)
-                        word.append(ch)
-                word = utils.to_unicode(b''.join(word), encoding=encoding, errors=unicode_errors)
-                weights = fromstring(fin.read(binary_len), dtype=REAL)
-                add_word(word, weights)
-        else:
-            for line_no in xrange(vocab_size):
-                line = fin.readline()
-                if line == b'':
-                    raise EOFError("unexpected end of input; is count incorrect or file otherwise damaged?")
-                parts = utils.to_unicode(line.rstrip(), encoding=encoding, errors=unicode_errors).split(" ")
-                if len(parts) != vector_size + 1:
-                    raise ValueError("invalid vector on line %s (is this really the text format?)" % line_no)
-                word, weights = parts[0], [REAL(x) for x in parts[1:]]
-                add_word(word, weights)
-    if result.vectors.shape[0] != len(result.vocab):
-        logger.info(
-            "duplicate words detected, shrinking matrix size from %i to %i",
-            result.vectors.shape[0], len(result.vocab)
-        )
-        result.vectors = ascontiguousarray(result.vectors[: len(result.vocab)])
-    assert (len(result.vocab), vector_size) == result.vectors.shape
-
-    logger.info("loaded %s matrix from %s", result.vectors.shape, fname)
-    return result
 
 
 # Example: ./word2vec.py -train data.txt -output vec.txt -size 200 -window 5 -sample 1e-4 \
